+++ tests/testthat/test-check_urban.R
@@ -1,3 +1,0 @@
-test_that("check_urban works", {
-  expect_true(check_urban())
-})