+++ NEWS.md
@@ -1,5 +1,8 @@
 # available (development version)
 
+* Removed Urban Dictionary support, as the udapi package is set to be
+  archived.
+
 # available 1.0.5
 
 * Gábor Csárdi is now the maintainer.